Extract Journey.Step to its own module
